requirements.txt
# Django==4.1.6
# djangorestframework==3.9.3
# psycopg2>=2.9.3,<2.10
# Pillow>=9.1.0,<9.2
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
django-model-utils==4.2.0
uwsgi>=2.0.20,<2.1
